Centralise logging config
