+++ openzt/src/resource_manager/mod_config.rs
@@ -4,80 +4,9 @@ use std::path::PathBuf;
 use std::sync::LazyLock;
 use std::sync::Mutex;
 use tracing::info;
-use tracing_subscriber::filter::LevelFilter;
 
-static LOGGING_INITIALIZED: LazyLock<Mutex<bool>> = LazyLock::new(|| Mutex::new(false));
-
-/// Initialize logging with settings from openzt.toml
-/// This should be called AFTER config is loaded
-pub fn init_logging(config: &LoggingConfig) -> anyhow::Result<()> {
-    let mut initialized = LOGGING_INITIALIZED.lock().unwrap();
-    if *initialized {
-        return Err(anyhow::anyhow!("Logging already initialized"));
-    }
-
-    let enable_ansi = enable_ansi_support::enable_ansi_support().is_ok();
-
-    #[cfg(not(feature = "integration-tests"))]
-    let level_filter = config.level.to_level_filter();
-    #[cfg(feature = "integration-tests")]
-    let level_filter = LevelFilter::TRACE; // Force TRACE level for integration tests
-
-    #[cfg(not(feature = "integration-tests"))]
-    let log_to_file = config.log_to_file;
-    #[cfg(feature = "integration-tests")]
-    let log_to_file = true;
-
-    use tracing_subscriber::layer::SubscriberExt;
-    use tracing_subscriber::util::SubscriberInitExt;
-    use tracing_subscriber::Layer;
-
-    // Always set up console logging
-    let console_layer = tracing_subscriber::fmt::layer()
-        .with_ansi(enable_ansi)
-        .with_writer(std::io::stdout)
-        .with_filter(level_filter);
-
-    // Set up file logging if enabled
-    if log_to_file {
-        let log_path = crate::util::get_base_path().join("openzt.log");
-        match std::fs::File::create(&log_path) {
-            Ok(log_file) => {
-                // Wrap in non-blocking writer
-                let (non_blocking, _guard) = tracing_appender::non_blocking(log_file);
-
-                let file_layer = tracing_subscriber::fmt::layer()
-                    .with_ansi(false) // No ANSI codes in file
-                    .with_writer(non_blocking)
-                    .with_filter(level_filter);
-
-                // Initialize with both console and file layers
-                tracing_subscriber::registry().with(console_layer).with(file_layer).init();
-
-                // Store guard to prevent it from being dropped
-                // Note: We need to leak this guard to keep file logging active
-                std::mem::forget(_guard);
-
-                eprintln!("Logging initialized: level={:?}, file={}", config.level, log_path.display());
-            }
-            Err(e) => {
-                // Fall back to console-only if file creation fails
-                tracing_subscriber::registry().with(console_layer).init();
-
-                eprintln!("Failed to create openzt.log: {}", e);
-                eprintln!("Logging initialized: level={:?}, console only", config.level);
-            }
-        }
-    } else {
-        // Console-only logging
-        tracing_subscriber::registry().with(console_layer).init();
-
-        eprintln!("Logging initialized: level={:?}, console only", config.level);
-    }
-
-    *initialized = true;
-    Ok(())
-}
+// Re-export LoggingConfig from the logging module
+pub use crate::logging::LoggingConfig;
 /// OpenZT configuration file structure (openzt.toml)
 #[derive(Deserialize, Serialize, Debug, Clone)]
 #[serde(deny_unknown_fields)]
@@ -125,45 +54,6 @@ pub struct ModLoadingConfig {
     /// Warn on conflicts (default: true)
     #[serde(default = "default_true")]
     pub warn_on_conflicts: bool,
-}
-
-/// Log level setting for OpenZT logging
-#[derive(Deserialize, Serialize, Debug, Clone, Copy, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-#[derive(Default)]
-pub enum LogLevel {
-    Trace,
-    Debug,
-    Info,
-    #[default]
-    Warn,
-    Error,
-}
-
-impl LogLevel {
-    /// Convert to tracing's LevelFilter
-    pub fn to_level_filter(self) -> LevelFilter {
-        match self {
-            LogLevel::Trace => LevelFilter::TRACE,
-            LogLevel::Debug => LevelFilter::DEBUG,
-            LogLevel::Info => LevelFilter::INFO,
-            LogLevel::Warn => LevelFilter::WARN,
-            LogLevel::Error => LevelFilter::ERROR,
-        }
-    }
-}
-
-/// Logging configuration section
-#[derive(Deserialize, Serialize, Debug, Clone)]
-#[serde(deny_unknown_fields)]
-pub struct LoggingConfig {
-    /// Enable file logging to openzt.log (default: true)
-    #[serde(default = "default_true")]
-    pub log_to_file: bool,
-
-    /// Log level (default: Warn)
-    #[serde(default)]
-    pub level: LogLevel,
 }
 
 /// Resource cache configuration section
@@ -241,10 +131,7 @@ impl Default for OpenZTConfig {
                 auto_resolve_new_mods: true,
                 warn_on_conflicts: true,
             },
-            logging: LoggingConfig {
-                log_to_file: true,
-                level: LogLevel::Warn,
-            },
+            logging: LoggingConfig::default(),
             resource_cache: ResourceCacheConfig::default(),
             expansions: ExpansionConfig::default(),
             dev: DevConfig::default(),
@@ -259,15 +146,6 @@ impl Default for ModLoadingConfig {
             disabled: Vec::new(),
             auto_resolve_new_mods: true,
             warn_on_conflicts: true,
-        }
-    }
-}
-
-impl Default for LoggingConfig {
-    fn default() -> Self {
-        LoggingConfig {
-            log_to_file: true,
-            level: LogLevel::Warn,
         }
     }
 }
@@ -508,6 +386,7 @@ fn get_temp_config_path() -> PathBuf {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::logging::LogLevel;
 
     #[test]
     fn test_default_config() {
